[Rename] 디렉토리명 변경 및 파일명 변경
