--- .vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-  "recommendations": [
-    "arcanis.vscode-zipfs",
-    "dbaeumer.vscode-eslint",
-    "esbenp.prettier-vscode"
-  ]
-}
