+++ .yarnrc.yml
@@ -2,4 +2,4 @@ nodeLinker: pnp
 packageExtensions:
   react-scripts@*:
     peerDependencies:
-      eslint-config-react-app: '*'
+      eslint-config-next-app: '*'
