+++ app/layout.tsx
@@ -3,10 +3,10 @@ import 'react-toastify/dist/ReactToastify.css';
 import { createServerSupabaseClient } from 'utils/supabase/server';
 import { ToastContainer } from 'react-toastify';
 import { Metadata } from 'next';
-import ReactQueryClientProvider from 'config/ReactQueryClientProvider';
-import dynamic from 'next/dynamic';
-import Auth from 'components/auth/common';
+import ReactQueryClientProvider from 'config/react-query-client-provider';
 import AuthProvider from 'config/auth-provider';
+import dynamic from 'next/dynamic';
+import Auth from 'components/auth/shared';
 import MainLayout from 'components/layouts/main-layout';
 import KakaoMap from 'components/layouts/kakaomap';
