+++ components/auth/shared/logo-image.tsx
